Move span logic to separate module
