+++ parser/src/lib.rs
@@ -55,237 +55,17 @@ extern crate alloc;
 
 pub use crate::{
     parser::{Error, SpannedError},
+    spans::{CodeFragment, InputSpan, LocatedSpan, MaybeSpanned, NomResult, Spanned, StripCode},
     traits::{Features, Grammar, GrammarExt},
 };
 
-use alloc::{borrow::ToOwned, boxed::Box, format, string::String, vec, vec::Vec};
+use alloc::{boxed::Box, vec, vec::Vec};
 use core::fmt;
-use nom::Slice;
 
 pub mod grammars;
-mod helpers;
 mod parser;
+mod spans;
 mod traits;
-
-/// Code span.
-pub type InputSpan<'a> = nom_locate::LocatedSpan<&'a str, ()>;
-/// Value with an associated code span.
-pub type Spanned<'a, T = ()> = LocatedSpan<&'a str, T>;
-/// Parsing outcome generalized by the type returned on success.
-pub type NomResult<'a, T> = nom::IResult<InputSpan<'a>, T, SpannedError<'a>>;
-
-// FIXME: move to sep. module
-/// FIXME
-#[derive(Debug, Clone, Copy)]
-pub struct LocatedSpan<Span, T = ()> {
-    offset: usize,
-    line: u32,
-    column: usize,
-    fragment: Span,
-
-    /// Extra information that can be embedded by the user.
-    pub extra: T,
-}
-
-impl<Span: PartialEq, T> PartialEq for LocatedSpan<Span, T> {
-    fn eq(&self, other: &Self) -> bool {
-        self.line == other.line && self.offset == other.offset && self.fragment == other.fragment
-    }
-}
-
-impl<Span, T> LocatedSpan<Span, T> {
-    /// The offset represents the position of the fragment relatively to
-    /// the input of the parser. It starts at offset 0.
-    pub fn location_offset(&self) -> usize {
-        self.offset
-    }
-
-    /// The line number of the fragment relatively to the input of the parser. It starts at line 1.
-    pub fn location_line(&self) -> u32 {
-        self.line
-    }
-
-    /// The column of the fragment start.
-    pub fn get_column(&self) -> usize {
-        self.column
-    }
-
-    /// The fragment that is spanned. The fragment represents a part of the input of the parser.
-    pub fn fragment(&self) -> &Span {
-        &self.fragment
-    }
-
-    /// FIXME
-    pub fn map_extra<U>(self, map_fn: impl FnOnce(T) -> U) -> LocatedSpan<Span, U> {
-        LocatedSpan {
-            offset: self.offset,
-            line: self.line,
-            column: self.column,
-            fragment: self.fragment,
-            extra: map_fn(self.extra),
-        }
-    }
-
-    /// FIXME
-    pub fn map_span<U>(self, map_fn: impl FnOnce(Span) -> U) -> LocatedSpan<U, T> {
-        LocatedSpan {
-            offset: self.offset,
-            line: self.line,
-            column: self.column,
-            fragment: map_fn(self.fragment),
-            extra: self.extra,
-        }
-    }
-}
-
-impl<'a, T> Spanned<'a, T> {
-    pub(crate) fn new(span: InputSpan<'a>, extra: T) -> Self {
-        Self {
-            offset: span.location_offset(),
-            line: span.location_line(),
-            column: span.get_column(),
-            fragment: *span.fragment(),
-            extra,
-        }
-    }
-}
-
-impl<'a> Spanned<'a> {
-    /// FIXME
-    pub fn from_str<R>(code: &'a str, range: R) -> Self
-    where
-        InputSpan<'a>: Slice<R>,
-    {
-        let input = InputSpan::new(code);
-        Self::new(input.slice(range), ())
-    }
-}
-
-impl<'a> MaybeSpanned<'a> {
-    /// FIXME
-    pub fn from_str<R>(code: &'a str, range: R) -> Self
-    where
-        InputSpan<'a>: Slice<R>,
-    {
-        Spanned::from_str(code, range).into()
-    }
-}
-
-impl<Span: Copy, T> LocatedSpan<Span, T> {
-    /// FIXME
-    pub fn copy_with_extra<U>(&self, value: U) -> LocatedSpan<Span, U> {
-        LocatedSpan {
-            offset: self.offset,
-            line: self.line,
-            column: self.column,
-            fragment: self.fragment,
-            extra: value,
-        }
-    }
-
-    /// FIXME
-    pub fn with_no_extra(&self) -> LocatedSpan<Span> {
-        self.copy_with_extra(())
-    }
-}
-
-impl<Span: Copy, T: Clone> LocatedSpan<Span, T> {
-    pub(crate) fn with_mapped_span<U>(&self, map_fn: impl FnOnce(Span) -> U) -> LocatedSpan<U, T> {
-        LocatedSpan {
-            offset: self.offset,
-            line: self.line,
-            column: self.column,
-            fragment: map_fn(self.fragment),
-            extra: self.extra.clone(),
-        }
-    }
-}
-
-impl<'a, T> From<nom_locate::LocatedSpan<&'a str, T>> for LocatedSpan<&'a str, T> {
-    fn from(value: nom_locate::LocatedSpan<&'a str, T>) -> Self {
-        Self {
-            offset: value.location_offset(),
-            line: value.location_line(),
-            column: value.get_column(),
-            fragment: *value.fragment(),
-            extra: value.extra,
-        }
-    }
-}
-
-/// FIXME
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub enum Code<'a> {
-    /// FIXME
-    Str(&'a str),
-    /// FIXME
-    Stripped(usize),
-}
-
-impl Code<'_> {
-    /// FIXME
-    pub fn strip(self) -> Code<'static> {
-        match self {
-            Self::Str(string) => Code::Stripped(string.len()),
-            Self::Stripped(len) => Code::Stripped(len),
-        }
-    }
-
-    /// FIXME
-    pub fn len(self) -> usize {
-        match self {
-            Self::Str(string) => string.len(),
-            Self::Stripped(len) => len,
-        }
-    }
-
-    /// FIXME
-    pub fn is_empty(self) -> bool {
-        self.len() == 0
-    }
-}
-
-impl<'a> From<&'a str> for Code<'a> {
-    fn from(value: &'a str) -> Self {
-        Code::Str(value)
-    }
-}
-
-/// Value with an optional associated code span.
-pub type MaybeSpanned<'a, T = ()> = LocatedSpan<Code<'a>, T>;
-
-impl<T> MaybeSpanned<'_, T> {
-    /// FIXME
-    pub fn code_or_location(&self, default_name: &str) -> String {
-        match self.fragment {
-            Code::Str(code) => code.to_owned(),
-            Code::Stripped(_) => format!("{} at {}:{}", default_name, self.line, self.column),
-        }
-    }
-}
-
-impl<'a, T> From<Spanned<'a, T>> for MaybeSpanned<'a, T> {
-    fn from(value: Spanned<'a, T>) -> Self {
-        value.map_span(Code::from)
-    }
-}
-
-/// FIXME
-pub trait StripCode {
-    /// FIXME
-    type Stripped: 'static;
-
-    /// FIXME
-    fn strip_code(&self) -> Self::Stripped;
-}
-
-impl<T: Clone + 'static> StripCode for MaybeSpanned<'_, T> {
-    type Stripped = MaybeSpanned<'static, T>;
-
-    fn strip_code(&self) -> Self::Stripped {
-        self.with_mapped_span(|code| code.strip())
-    }
-}
 
 /// Parsing context.
 // TODO: Add more fine-grained contexts.
